FIX Slugged headers now work locally
